Export hook k
