+++ .babelrc
@@ -12,7 +12,8 @@
   "plugins": [
     "@babel/plugin-transform-flow-strip-types",
     "@babel/plugin-transform-spread",
-    "@babel/plugin-proposal-class-properties"
+    "@babel/plugin-proposal-class-properties",
+    "@babel/plugin-proposal-object-rest-spread"
   ],
   "env": {
     "next": {
